// apps/application-dashboard/src/deprecated/apis/paymentApi.types.ts
export type TemplateDetailsType = {
  id: string;
  name: string;
  details: {
    order: string;
    source_account: AccountDetailsType;
    destination_account: AccountDetailsType;
    recipient_name: string;
    recipient_id: string;
  }[];
  created_by: string;
  creation_timestamp: string;
  type: MOVE_MONEY_TYPE;
  status?: STATUS_TYPES;
  can_approve: boolean;
  policy_result_id?: string;
  approval_id?: string;
};